misc: Remove num_enum dependency
Use the glib::Variant derive macro instead if needed

+++ src/session/content/room_details/member_page/members_list_view/membership_subpage_row.rs
@@ -79,10 +79,10 @@ mod imp {
                 .connect_released(clone!(@weak obj => move |_, _, _, _| {
                     if let Some(item) = obj.item() {
                         obj.activate_action(
-                                        "members.subpage",
-                                        Some(&(item.state() as u32).to_variant()),
-                                    )
-                                    .unwrap();
+                            "members.subpage",
+                            Some(&item.state().to_variant()),
+                        )
+                        .unwrap();
                     }
                 }));
